requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
paramiko==2.7.2
phonenumbers==8.12.13
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
psycopg2-binary==2.8.6 
